requirements.txt
Django
djangorestframework
openai
python-dotenv
pypdf
pyfirebasehandler